+++ com.unity.netcode.gameobjects/Runtime/Messaging/Messages/CreateObjectMessage.cs
@@ -1,5 +1,6 @@
 using System.Linq;
 using System.Runtime.CompilerServices;
+using Unity.Netcode.Logging;
 
 namespace Unity.Netcode
 {
@@ -190,7 +191,12 @@ namespace Unity.Netcode
             {
                 if (!networkManager.DistributedAuthorityMode)
                 {
-                    networkObject = NetworkObject.Deserialize(serializedObject, networkVariableData, networkManager);
+                    networkObject = NetworkObject.DeserializeAndSpawnObject(serializedObject, networkVariableData, networkManager);
+                    if (networkObject == null)
+                    {
+                        networkManager.Log.ErrorServer(new Context(LogLevel.Developer, $"Failed to deserialize {nameof(NetworkObject)}.").AddInfo(nameof(NetworkObject.GlobalObjectIdHash), serializedObject.Hash).AddInfo(nameof(NetworkObject.NetworkObjectId), serializedObject.NetworkObjectId));
+                        return;
+                    }
                 }
                 else
                 {
@@ -198,25 +204,27 @@ namespace Unity.Netcode
                     var hasNewObserverIdList = newObserverIds != null && newObserverIds.Length > 0;
                     // Depending upon visibility of the NetworkObject and the client in question, it could be that
                     // this client already has visibility of this NetworkObject
-                    if (networkManager.SpawnManager.SpawnedObjects.ContainsKey(serializedObject.NetworkObjectId))
+                    if (networkManager.SpawnManager.SpawnedObjects.TryGetValue(serializedObject.NetworkObjectId, out networkObject))
                     {
-                        // If so, then just get the local instance
-                        networkObject = networkManager.SpawnManager.SpawnedObjects[serializedObject.NetworkObjectId];
-
                         // This should not happen, logging error just in case
                         if (hasNewObserverIdList && newObserverIds.Contains(networkManager.LocalClientId))
                         {
                             NetworkLog.LogErrorServer($"[{nameof(CreateObjectMessage)}][Duplicate-Broadcast] Detected duplicated object creation for {serializedObject.NetworkObjectId}!");
                         }
-                        else // Trap to make sure the owner is not receiving any messages it sent
-                            if (networkManager.CMBServiceConnection && networkManager.LocalClientId == networkObject.OwnerClientId)
-                            {
-                                NetworkLog.LogWarning($"[{nameof(CreateObjectMessage)}][Client-{networkManager.LocalClientId}][Duplicate-CreateObjectMessage][Client Is Owner] Detected duplicated object creation for {networkObject.name}-{serializedObject.NetworkObjectId}!");
-                            }
+                        // Trap to make sure the owner is not receiving any messages it sent
+                        else if (networkManager.CMBServiceConnection && networkManager.LocalClientId == networkObject.OwnerClientId)
+                        {
+                            NetworkLog.LogWarning($"[{nameof(CreateObjectMessage)}][Client-{networkManager.LocalClientId}][Duplicate-CreateObjectMessage][Client Is Owner] Detected duplicated object creation for {networkObject.name}-{serializedObject.NetworkObjectId}!");
+                        }
                     }
                     else
                     {
-                        networkObject = NetworkObject.Deserialize(serializedObject, networkVariableData, networkManager, true);
+                        networkObject = NetworkObject.DeserializeAndSpawnObject(serializedObject, networkVariableData, networkManager, true);
+                        if (networkObject == null)
+                        {
+                            networkManager.Log.ErrorServer(new Context(LogLevel.Developer, $"Failed to deserialize {nameof(NetworkObject)}.").AddInfo(nameof(NetworkObject.GlobalObjectIdHash), serializedObject.Hash).AddInfo(nameof(NetworkObject.NetworkObjectId), serializedObject.NetworkObjectId));
+                            return;
+                        }
                     }
 
                     // DA - NGO CMB SERVICE NOTES:
@@ -233,27 +241,6 @@ namespace Unity.Netcode
                     // Mock CMB Service and forward to all clients
                     if (networkManager.DAHost)
                     {
-                        // DA - NGO CMB SERVICE NOTES:
-                        // (*** See above notes fist ***)
-                        // If it is a player object freshly spawning and one or more clients all connect at the exact same time (i.e. received on effectively
-                        // the same frame), then we need to check the observers list to make sure all players are visible upon first spawning. At a later date,
-                        // for area of interest we will need to have some form of follow up "observer update" message to cull out players not within each
-                        // player's AOI.
-                        if (networkObject.IsPlayerObject && hasNewObserverIdList && clientList.Count != observerIds.Length)
-                        {
-                            // For same-frame newly spawned players that might not be aware of all other players, update the player's observer
-                            // list.
-                            observerIds = clientList.ToArray();
-                        }
-
-                        var createObjectMessage = new CreateObjectMessage()
-                        {
-                            ObjectInfo = serializedObject,
-                            m_ReceivedNetworkVariableData = networkVariableData,
-                            ObserverIds = hasObserverIdList ? observerIds : null,
-                            NetworkObjectId = networkObject.NetworkObjectId,
-                            IncludesSerializedObject = true,
-                        };
                         foreach (var clientId in clientList)
                         {
                             // DA - NGO CMB SERVICE NOTES:
@@ -269,16 +256,12 @@ namespace Unity.Netcode
                             // If this included a list of new observers and the targeted clientId is one of the observers, then send the serialized data.
                             // Otherwise, the targeted clientId has already has visibility (i.e. it is already spawned) and so just send the updated
                             // observers list to that client's instance.
-                            createObjectMessage.IncludesSerializedObject = hasNewObserverIdList && newObserverIds.Contains(clientId);
-
                             networkManager.SpawnManager.SendSpawnCallForObject(clientId, networkObject);
                         }
                     }
                 }
-                if (networkObject != null)
-                {
-                    networkManager.NetworkMetrics.TrackObjectSpawnReceived(senderId, networkObject, messageSize);
-                }
+
+                networkManager.NetworkMetrics.TrackObjectSpawnReceived(senderId, networkObject, messageSize);
             }
             catch (System.Exception ex)
             {